+++ heatmiserneo/CHANGELOG.md
@@ -3,3 +3,5 @@
 ## 3.0.0
 - It is supported to be added through UI.
 - Related actions has the fields configured for easy action.
+- Reconfigure support.
+- Force update to the thermostats on any change to thermostat via the integration.